pyproject.toml
[tool.poetry]
name = "source-data"
version = "0.1.0"
description = "App for validating and ingesting source data in the InTaVia Knowledge graph (IKG)"
authors = ["Matthias Schlögl <matthias.schloegl@oeaw.ac.at>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
rdflib = "^6.3.2"
pyshacl = "^0.22.2"
requests = "^2.29.0"
SPARQLWrapper = "^2.0.0"
click = "^8.1.3"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
